Begin a complete refactor
