+++ abi/src/main/java/org/web3j/abi/datatypes/CompositeDataHolder.java
@@ -18,11 +18,11 @@ import java.util.List;
 
 public class CompositeDataHolder extends DynamicStruct {
 
-    public List<Class<Type>> getTypes() {
+    public static List<Class<Type>> getTypes() {
         return itemTypes;
     }
 
-    private final List<Class<Type>> itemTypes = new ArrayList<>();
+    private static final List<Class<Type>> itemTypes = new ArrayList<>();
 
     public CompositeDataHolder(List<Type> values) {
         super(values);
