+++ codegen/src/main/java/org/web3j/codegen/GenerationReporter.java
@@ -13,6 +13,6 @@
 package org.web3j.codegen;
 
 /** Can be used to provide report about a code generation process. */
-interface GenerationReporter {
+public interface GenerationReporter {
     void report(String msg);
 }
