+++ codegen/src/main/java/org/web3j/codegen/unit/wrapper/generators/ConstructorGenerator.java
@@ -1,0 +1,61 @@
+/*
+ * Copyright 2020 Web3 Labs Ltd.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+package org.web3j.codegen.unit.wrapper.generators;
+
+import java.math.BigInteger;
+import javax.lang.model.element.Modifier;
+
+import com.squareup.javapoet.MethodSpec;
+
+import org.web3j.protocol.Web3j;
+import org.web3j.tx.gas.ContractGasProvider;
+
+import static org.web3j.codegen.unit.wrapper.Constants.*;
+
+public class ConstructorGenerator {
+
+    public static MethodSpec buildConstructor(
+            Class authType, String authName, boolean withGasProvider) {
+        MethodSpec.Builder toReturn =
+                MethodSpec.constructorBuilder()
+                        .addModifiers(Modifier.PROTECTED)
+                        .addParameter(String.class, CONTRACT_ADDRESS)
+                        .addParameter(Web3j.class, WEB3J)
+                        .addParameter(authType, authName);
+
+        if (withGasProvider) {
+            toReturn.addParameter(ContractGasProvider.class, CONTRACT_GAS_PROVIDER)
+                    .addStatement(
+                            "super($N, $N, $N, $N, $N)",
+                            BINARY,
+                            CONTRACT_ADDRESS,
+                            WEB3J,
+                            authName,
+                            CONTRACT_GAS_PROVIDER);
+        } else {
+            toReturn.addParameter(BigInteger.class, GAS_PRICE)
+                    .addParameter(BigInteger.class, GAS_LIMIT)
+                    .addStatement(
+                            "super($N, $N, $N, $N, $N, $N)",
+                            BINARY,
+                            CONTRACT_ADDRESS,
+                            WEB3J,
+                            authName,
+                            GAS_PRICE,
+                            GAS_LIMIT)
+                    .addAnnotation(Deprecated.class);
+        }
+
+        return toReturn.build();
+    }
+}
